Balancer interface updates
git-svn-id: https://ha-jdbc.svn.sourceforge.net/svnroot/ha-jdbc/trunk@933 882a69a1-a10d-0410-8702-ae3607fe02de

+++ test/net/sf/hajdbc/balancer/TestRandomBalancer.java
@@ -46,7 +46,7 @@ public class TestRandomBalancer extends AbstractTestBalancer
 		
 		for (int i = 0; i < count; ++i)
 		{
-			results[balancer.next().getWeight().intValue()] += 1;
+			results[balancer.next().getWeight()] += 1;
 		}
 
 		assertTrue(results[0] == count);
@@ -59,7 +59,7 @@ public class TestRandomBalancer extends AbstractTestBalancer
 		
 		for (int i = 0; i < count; ++i)
 		{
-			results[balancer.next().getWeight().intValue()] += 1;
+			results[balancer.next().getWeight()] += 1;
 		}
 		
 		assertTrue(results[0] == 0);
@@ -72,7 +72,7 @@ public class TestRandomBalancer extends AbstractTestBalancer
 		
 		for (int i = 0; i < count; ++i)
 		{
-			results[balancer.next().getWeight().intValue()] += 1;
+			results[balancer.next().getWeight()] += 1;
 		}
 
 		assertTrue(results[0] == 0);
